backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
pydantic==2.10.6
python-dotenv==1.0.0
google-generativeai==0.3.1
httpx==0.27.0 
python-multipart==0.0.9
youtube-transcript-api==0.6.1
Pillow
matplotlib
openai 